fix last admin delete, fix 404 routing

+++ src/main/java/ua/softserve/rv_028/issuecitymonitor/service/UserService.java
@@ -59,8 +59,12 @@ public class UserService {
     }
 
     public void deleteById(long id) {
-        userDao.delete(id);
-        LOGGER.debug("Deleted user " + id);
+        if (userDao.countAdmins() <= 1 && (findOne(id).getUserRole() == UserRole.ADMIN)) {
+            throw new LastAdminException();
+        } else {
+            userDao.delete(id);
+            LOGGER.debug("Deleted user " + id);
+        }
     }
 
     private User findOne(long id){
